requirements.txt
pillow
numpy
opencv-python
torch
torchvision
tensorflow
tqdm
